fix: resolve compilation errors for Unity 2022 types
- Create ComponentAttributes.cs with RequireComponent, AddComponentMenu, etc.
- Remove duplicate type definitions (Coroutine, AnimatorStateInfo, AnimatorClipInfo, etc.)
- Fix Mathf.Infinity usage (replace with float.PositiveInfinity)
- Fix type conversion errors in AudioSource and WheelCollider
- Build succeeds with 0 errors

+++ anity-lib-core/src/Anity.Core/UnityCompat/Runtime/AudioSource.cs
@@ -83,7 +83,7 @@ public class AudioSource : Behaviour
     public int priority
     {
         get => _priority;
-        set => _priority = Mathf.Clamp(value, 0, 256);
+        set => _priority = (int)Mathf.Clamp(value, 0, 256);
     }
 
     public float minDistance
